02/assignment/waffleblog/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.5
defusedxml==0.7.1
Django==4.2.7
django-allauth==0.58.2
django-braces==1.15.0
idna==3.4
mysql==0.0.3
mysqlclient==2.2.0
oauthlib==3.2.2
pycparser==2.21
PyJWT==2.8.0
PyMySQL==1.1.0
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
urllib3==2.0.7
